Eliminate test flakiness + standardize test setup and assertions
- Import and call resetLoggedEvents() before each test to reset shared state
- Replace expect() assertions with expect.poll() for async events
- Remove unused auth test events and helpers
