+++ vscode/test/e2e/auth.test.ts
@@ -1,25 +1,15 @@
 import { expect } from '@playwright/test'
 
-import { loggedEvents, SERVER_URL, VALID_TOKEN } from '../fixtures/mock-server'
+import { loggedEvents, resetLoggedEvents, SERVER_URL, VALID_TOKEN } from '../fixtures/mock-server'
 
 import { signOut, test } from './helpers'
 
-const expectedOrderedEvents = [
-    'CodyInstalled',
-    'CodyVSCodeExtension:Auth:failed',
-    'CodyVSCodeExtension:auth:clickOtherSignInOptions',
-    'CodyVSCodeExtension:login:clicked',
-    'CodyVSCodeExtension:auth:selectSigninMenu',
-    'CodyVSCodeExtension:auth:fromToken',
-    'CodyVSCodeExtension:Auth:failed',
-    'CodyVSCodeExtension:auth:clickOtherSignInOptions',
-    'CodyVSCodeExtension:login:clicked',
-    'CodyVSCodeExtension:auth:selectSigninMenu',
-    'CodyVSCodeExtension:auth:fromToken',
-    'CodyVSCodeExtension:Auth:connected',
-    'CodyVSCodeExtension:logout:clicked',
-    'CodyVSCodeExtension:Auth:disconnected',
-]
+const expectedOrderedEvents = ['CodyVSCodeExtension:logout:clicked', 'CodyVSCodeExtension:Auth:disconnected']
+
+test.beforeEach(() => {
+    void resetLoggedEvents()
+})
+
 test('requires a valid auth token and allows logouts', async ({ page, sidebar }) => {
     await expect(sidebar.getByText('Invalid credentials')).not.toBeVisible()
     await sidebar.getByRole('button', { name: 'Other Sign In Options…' }).click()
@@ -51,5 +41,5 @@ test('requires a valid auth token and allows logouts', async ({ page, sidebar })
 
     await expect(sidebar.getByRole('button', { name: 'Other Sign In Options…' })).toBeVisible()
     await expect(sidebar.getByText('Invalid credentials')).not.toBeVisible()
-    expect(loggedEvents).toEqual(expectedOrderedEvents)
+    await expect.poll(() => loggedEvents).toEqual(expectedOrderedEvents)
 })
