+++ vscode/test/e2e/inline-assist-inline-touch.test.ts
@@ -1,18 +1,11 @@
 import { expect } from '@playwright/test'
 
-import { loggedEvents } from '../fixtures/mock-server'
+import { loggedEvents, resetLoggedEvents } from '../fixtures/mock-server'
 
 import { sidebarExplorer, sidebarSignin } from './common'
 import { test } from './helpers'
 
 const expectedOrderedEvents = [
-    'CodyInstalled',
-    'CodyVSCodeExtension:Auth:failed',
-    'CodyVSCodeExtension:auth:clickOtherSignInOptions',
-    'CodyVSCodeExtension:login:clicked',
-    'CodyVSCodeExtension:auth:selectSigninMenu',
-    'CodyVSCodeExtension:auth:fromToken',
-    'CodyVSCodeExtension:Auth:connected',
     'CodyVSCodeExtension:fixup:created',
     'CodyVSCodeExtension:keywordContext:searchDuration',
     'CodyVSCodeExtension:recipe:fixup:executed',
@@ -21,6 +14,9 @@ const expectedOrderedEvents = [
     'CodyVSCodeExtension:fixup:codeLens:clicked',
     'CodyVSCodeExtension:fixup:applied',
 ]
+test.beforeEach(() => {
+    resetLoggedEvents()
+})
 test.skip('start a fixup job from inline chat with valid auth', async ({ page, sidebar }) => {
     // Sign into Cody
     await sidebarSignin(page, sidebar)
@@ -53,7 +49,7 @@ test.skip('start a fixup job from inline chat with valid auth', async ({ page, s
 
     // Check if a new file called index.cody.html is created
     await expect(page.getByText('index.cody.html')).toBeVisible()
-    expect(loggedEvents).toEqual(expectedOrderedEvents)
+    await expect.poll(() => loggedEvents).toEqual(expectedOrderedEvents)
 
     // TODO check if content is correct. Currently blocked by ability to highlight in test
 })
