+++ vscode/test/e2e/inline-assist.test.ts
@@ -1,18 +1,11 @@
 import { expect } from '@playwright/test'
 
-import { loggedEvents } from '../fixtures/mock-server'
+import { loggedEvents, resetLoggedEvents } from '../fixtures/mock-server'
 
 import { sidebarExplorer, sidebarSignin } from './common'
 import { test } from './helpers'
 
-const expectedOrderedEvent = [
-    'CodyInstalled',
-    'CodyVSCodeExtension:Auth:failed',
-    'CodyVSCodeExtension:auth:clickOtherSignInOptions',
-    'CodyVSCodeExtension:login:clicked',
-    'CodyVSCodeExtension:auth:selectSigninMenu',
-    'CodyVSCodeExtension:auth:fromToken',
-    'CodyVSCodeExtension:Auth:connected',
+const expectedOrderedEvents = [
     'CodyVSCodeExtension:fixup:created',
     'CodyVSCodeExtension:keywordContext:searchDuration',
     'CodyVSCodeExtension:recipe:fixup:executed',
@@ -21,6 +14,10 @@ const expectedOrderedEvent = [
     'CodyVSCodeExtension:fixup:codeLens:clicked',
     'CodyVSCodeExtension:fixup:applied',
 ]
+
+test.beforeEach(() => {
+    resetLoggedEvents()
+})
 test('start a fixup job from inline chat with valid auth', async ({ page, sidebar }) => {
     // Sign into Cody
     await sidebarSignin(page, sidebar)
@@ -52,5 +49,5 @@ test('start a fixup job from inline chat with valid auth', async ({ page, sideba
     await expect(page.getByText('Fixup ready')).toBeVisible()
     await page.getByRole('button', { name: 'Apply' }).click()
     await expect(page.getByText('<title>Goodbye Cody</title>')).toBeVisible()
-    expect(loggedEvents).toEqual(expectedOrderedEvent)
+    await expect.poll(() => loggedEvents).toEqual(expectedOrderedEvents)
 })
